redone fish configs

+++ init.sh
@@ -1,29 +1,29 @@
 #!/bin/sh
 
-#ln -s ~/config/.zprofile ~
-#ln -s ~/config/.tmux.conf ~
-ln -s ~/config/.vimrc ~
-#ln -s ~/config/.zshrc ~
-ln -s ~/config/.ghci ~
-ln -s ~/config/.gitconfig ~
-#ln -s ~/config/.profile ~
+# Symlinks in home directory
+[ -e ~/.vimrc ] || ln -s ~/config/.vimrc ~
+[ -e ~/.ghci ] || ln -s ~/config/.ghci ~
+[ -e ~/.gitconfig ] || ln -s ~/config/.gitconfig ~
 
-mkdir ~/.config
+# Create .config if it doesn't exist
+[ -d ~/.config ] || mkdir ~/.config
 
-mkdir ~/.config/helix
-ln -s ~/config/.config/helix/config.toml ~/.config/helix
-ln -s ~/config/.config/helix/languages.toml ~/.config/helix
+# Helix setup
+[ -d ~/.config/helix ] || mkdir ~/.config/helix
+[ -e ~/.config/helix/config.toml ] || ln -s ~/config/.config/helix/config.toml ~/.config/helix
+[ -e ~/.config/helix/languages.toml ] || ln -s ~/config/.config/helix/languages.toml ~/.config/helix
 
-mkdir ~/.config/kitty
-ln -s ~/config/.config/kitty/theme.conf ~/.config/kitty
-ln -s ~/config/.config/kitty/kitty.conf ~/.config/kitty
-ln -s ~/config/.config/kitty/kitty.app.png ~/.config/kitty
-ln -s ~/config/.config/kitty/afterglow.conf ~/.config/kitty
+# Kitty setup
+[ -d ~/.config/kitty ] || mkdir ~/.config/kitty
+[ -e ~/.config/kitty/theme.conf ] || ln -s ~/config/.config/kitty/theme.conf ~/.config/kitty
+[ -e ~/.config/kitty/kitty.conf ] || ln -s ~/config/.config/kitty/kitty.conf ~/.config/kitty
+[ -e ~/.config/kitty/kitty.app.png ] || ln -s ~/config/.config/kitty/kitty.app.png ~/.config/kitty
+[ -e ~/.config/kitty/afterglow.conf ] || ln -s ~/config/.config/kitty/afterglow.conf ~/.config/kitty
 
-mkdir ~/.config/fish
-ln -s ~/config/.config/fish/config.fish ~/.config/fish
-ln -s ~/config/.config/fish/fish_plugins ~/.config/fish
+# Fish symlink (not directory creation)
+[ -d ~/.config/fish ] || ln -s ~/config/.config/fish ~/.config/fish
 
-mkdir ~/.config/zed
-ln -s ~/config/.config/zed/settings.json ~/.config/zed
-ln -s ~/config/.config/zed/keymap.json ~/.config/zed
+# Zed setup
+[ -d ~/.config/zed ] || mkdir ~/.config/zed
+[ -e ~/.config/zed/settings.json ] || ln -s ~/config/.config/zed/settings.json ~/.config/zed
+[ -e ~/.config/zed/keymap.json ] || ln -s ~/config/.config/zed/keymap.json ~/.config/zed
